Minor Codefixes and README part added
Unnecessary Codebits in bcd.py and test_bcd.py were removed.

A bug in the BcdDataset __init__ function call was removed. The default
"name" value did not match with the entry in the datasets.jason file.

The right bcd.csv file was added.

A bcd.md and the bcd part in the README.md was added.

The test_bcd.py file was expandet with a "load_all_in_mem" test. Also
the mock image size was altered to the right size.